Taxa de pagamento

+++ models/produtoQtde.py
@@ -2,4 +2,4 @@ from pydantic import BaseModel
 
 class produtoQtde(BaseModel):
     DESCRICAO_PRODUTO: str
-    QTDE: int
+    QTDE: float
